Add async job support.
This allows us to run async code on the worker thread.

A two-stage approach is used: first, a `Send` function is passed to
the thread, which is executed to obtain a non-`Send` future.  This
supports the ‘work-dealing’ pattern in which the task is first moved
to another thread, and then kept there whilst running.